Rename definitions module to templates.
